+++ admin/template/user.php
@@ -71,7 +71,7 @@ if ($page1 == "e" || $page1 == "edp" || $page1 == "ene") { ?>
 			<div class="col-md-6">
 				<div class="input-group">
 					<select name="jak_group" class="form-control selectpicker" data-size="5">
-						<?php if (isset($JAK_USERGROUP_ALL) && is_array($JAK_USERGROUP_ALL)) foreach ($JAK_USERGROUP_ALL as $z) {
+						<?php if (isset($JAK_USERGROUP_ALL) && is_array ($JAK_USERGROUP_ALL)) foreach ($JAK_USERGROUP_ALL as $z) {
 							if ($z["id"] != 1) { ?>
 								<option value="<?php echo $z["id"]; ?>"><?php echo $z["name"]; ?></option><?php }
 						} ?>
@@ -85,7 +85,7 @@ if ($page1 == "e" || $page1 == "edp" || $page1 == "ene") { ?>
 
 	<hr>
 
-<?php if (isset($JAK_USER_ALL_APPROVE) && is_array($JAK_USER_ALL_APPROVE)) { ?>
+<?php if (isset($JAK_USER_ALL_APPROVE) && is_array ($JAK_USER_ALL_APPROVE)) { ?>
 
 	<h3><?php echo $tl["user_box_table"]["usertb5"]; ?></h3>
 
@@ -107,15 +107,13 @@ if ($page1 == "e" || $page1 == "edp" || $page1 == "ene") { ?>
 						<th><?php echo $tl["user_box_table"]["usertb2"]; ?></th>
 						<th></th>
 						<th>
-							<button type="submit" name="password" id="button_key" class="btn btn-default btn-xs"
-											onclick="if(!confirm('<?php echo $tl["user_notification"]["pass1"]; ?>'))return false;">
+							<button type="submit" name="password" id="button_key" class="btn btn-default btn-xs" onclick="if(!confirm('<?php echo $tl["user_notification"]["pass1"]; ?>'))return false;">
 								<i class="fa fa-key"></i>
 							</button>
 						</th>
 						<th></th>
 						<th>
-							<button type="submit" name="delete" id="button_delete" class="btn btn-danger btn-xs"
-											onclick="if(!confirm('<?php echo $tl["user_notification"]["delall"]; ?>'))return false;">
+							<button type="submit" name="delete" id="button_delete" class="btn btn-danger btn-xs" onclick="if(!confirm('<?php echo $tl["user_notification"]["delall"]; ?>'))return false;">
 								<i class="fa fa-trash-o"></i>
 							</button>
 						</th>
@@ -126,35 +124,29 @@ if ($page1 == "e" || $page1 == "edp" || $page1 == "ene") { ?>
 							<td><?php echo $va["id"]; ?></td>
 							<td>
 								<div class="checkbox-singel check-success">
-									<input type="checkbox" id="jak_delete_user<?php echo $va["id"]; ?>" name="jak_delete_user[]"
-												 class="highlight" value="<?php echo $va["id"]; ?>"/>
+									<input type="checkbox" id="jak_delete_user<?php echo $va["id"]; ?>" name="jak_delete_user[]" class="highlight" value="<?php echo $va["id"]; ?>"/>
 									<label for="jak_delete_user<?php echo $va["id"]; ?>"></label>
 								</div>
 							</td>
 							<td>
-								<a
-									href="index.php?p=user&amp;sp=edit&amp;ssp=<?php echo $va["id"]; ?>"><?php echo $va["username"]; ?></a>
+								<a href="index.php?p=user&amp;sp=edit&amp;ssp=<?php echo $va["id"]; ?>"><?php echo $va["username"]; ?></a>
 							</td>
 							<td>
 								<a href="index.php?p=user&amp;sp=edit&amp;ssp=<?php echo $va["id"]; ?>"><?php echo $va["email"]; ?></a>
 							</td>
 							<td>
-								<?php if (isset($JAK_USERGROUP_ALL) && is_array($JAK_USERGROUP_ALL)) foreach ($JAK_USERGROUP_ALL as $z) {
-									if ($va["usergroupid"] == $z["id"]) { ?><a
-										href="index.php?p=usergroup&amp;sp=user&amp;ssp=<?php echo $z["id"]; ?>"><?php echo $z["name"]; ?></a><?php }
+								<?php if (isset($JAK_USERGROUP_ALL) && is_array ($JAK_USERGROUP_ALL)) foreach ($JAK_USERGROUP_ALL as $z) {
+									if ($va["usergroupid"] == $z["id"]) { ?>
+										<a										href="index.php?p=usergroup&amp;sp=user&amp;ssp=<?php echo $z["id"]; ?>"><?php echo $z["name"]; ?></a><?php }
 								} ?>
 							</td>
 							<td class="content-go">
-								<a href="index.php?p=user&amp;sp=verify&amp;ssp=<?php echo $va["id"]; ?>"
-									 class="btn btn-default btn-xs">
-									<i
-										class="fa fa-<?php if ($va["access"] == 3 || $va["access"] == 2) { ?>envelope-o<?php } else { ?>lock<?php } ?>"></i>
+								<a href="index.php?p=user&amp;sp=verify&amp;ssp=<?php echo $va["id"]; ?>" class="btn btn-default btn-xs">
+									<i class="fa fa-<?php if ($va["access"] == 3 || $va["access"] == 2) { ?>envelope-o<?php } else { ?>lock<?php } ?>"></i>
 								</a>
 							</td>
 							<td>
-								<a class="btn btn-default btn-xs"
-									 href="index.php?p=user&amp;sp=password&amp;ssp=<?php echo $va["id"]; ?>"
-									 onclick="if(!confirm('<?php echo $tl["user_notification"]["pass"]; ?>'))return false;">
+								<a class="btn btn-default btn-xs" href="index.php?p=user&amp;sp=password&amp;ssp=<?php echo $va["id"]; ?>" onclick="if(!confirm('<?php echo $tl["user_notification"]["pass"]; ?>'))return false;">
 									<i class="fa fa-key"></i>
 								</a>
 							</td>
@@ -164,9 +156,7 @@ if ($page1 == "e" || $page1 == "edp" || $page1 == "ene") { ?>
 								</a>
 							</td>
 							<td>
-								<a class="btn btn-default btn-xs" href="index.php?p=user&amp;sp=delete&amp;ssp=<?php echo $va["id"]; ?>"
-									 data-confirm="<?php echo $tl["user_notification"]["del"]; ?>" data-toggle="tooltip"
-									 data-placement="bottom" title="<?php echo $tl["icons"]["i1"]; ?>" )return false;">
+								<a class="btn btn-default btn-xs" href="index.php?p=user&amp;sp=delete&amp;ssp=<?php echo $va["id"]; ?>" data-confirm="<?php echo $tl["user_notification"]["del"]; ?>" data-toggle="tooltip" data-placement="bottom" title="<?php echo $tl["icons"]["i1"]; ?>" )return false;">
 								<i class="fa fa-trash-o"></i>
 								</a>
 							</td>
@@ -220,27 +210,24 @@ if ($page1 == "e" || $page1 == "edp" || $page1 == "ene") { ?>
 							</button>
 						</th>
 						<th>
-							<button type="submit" name="password" id="button_key" class="btn btn-default btn-xs"
-											onclick="if(!confirm('<?php echo $tl["user_notification"]["pass1"]; ?>'))return false;">
+							<button type="submit" name="password" id="button_key" class="btn btn-default btn-xs" onclick="if(!confirm('<?php echo $tl["user_notification"]["pass1"]; ?>'))return false;">
 								<i class="fa fa-key"></i>
 							</button>
 						</th>
 						<th></th>
 						<th>
-							<button type="submit" name="delete" id="button_delete" class="btn btn-danger btn-xs"
-											data-confirm-del="<?php echo $tl["user_notification"]["delall"]; ?>" disabled="disabled">
+							<button type="submit" name="delete" id="button_delete" class="btn btn-danger btn-xs" data-confirm-del="<?php echo $tl["user_notification"]["delall"]; ?>" disabled="disabled">
 								<i class="fa fa-trash-o"></i>
 							</button>
 						</th>
 					</tr>
 					</thead>
-					<?php if (isset($JAK_USER_ALL) && is_array($JAK_USER_ALL)) foreach ($JAK_USER_ALL as $v) { ?>
+					<?php if (isset($JAK_USER_ALL) && is_array ($JAK_USER_ALL)) foreach ($JAK_USER_ALL as $v) { ?>
 						<tr>
 							<td><?php echo $v["id"]; ?></td>
 							<td>
 								<div class="checkbox-singel check-success">
-									<input type="checkbox" id="jak_delete_user<?php echo $v["id"]; ?>" name="jak_delete_user[]"
-												 class="highlight" value="<?php echo $v["id"]; ?>"/>
+									<input type="checkbox" id="jak_delete_user<?php echo $v["id"]; ?>" name="jak_delete_user[]" class="highlight" value="<?php echo $v["id"]; ?>"/>
 									<label for="jak_delete_user<?php echo $v["id"]; ?>"></label>
 								</div>
 							</td>
@@ -251,10 +238,9 @@ if ($page1 == "e" || $page1 == "edp" || $page1 == "ene") { ?>
 								<a href="index.php?p=user&amp;sp=edit&amp;ssp=<?php echo $v["id"]; ?>"><?php echo $v["email"]; ?></a>
 							</td>
 							<td>
-								<?php if (isset($JAK_USERGROUP_ALL) && is_array($JAK_USERGROUP_ALL)) foreach ($JAK_USERGROUP_ALL as $z) {
+								<?php if (isset($JAK_USERGROUP_ALL) && is_array ($JAK_USERGROUP_ALL)) foreach ($JAK_USERGROUP_ALL as $z) {
 									if ($v["usergroupid"] == $z["id"]) { ?>
-										<a
-											href="index.php?p=usergroup&amp;sp=user&amp;ssp=<?php echo $z["id"]; ?>"><?php echo $z["name"]; ?></a>
+										<a href="index.php?p=usergroup&amp;sp=user&amp;ssp=<?php echo $z["id"]; ?>"><?php echo $z["name"]; ?></a>
 									<?php }
 								} ?>
 							</td>
@@ -268,8 +254,7 @@ if ($page1 == "e" || $page1 == "edp" || $page1 == "ene") { ?>
 								?>
 							</td>
 							<td>
-								<a class="btn btn-default btn-xs" href="index.php?p=user&amp;sp=lock&amp;ssp=<?php echo $v["id"]; ?>"
-									 data-toggle="tooltip" data-placement="bottom" title="<?php if ($v["access"] == '1') {
+								<a class="btn btn-default btn-xs" href="index.php?p=user&amp;sp=lock&amp;ssp=<?php echo $v["id"]; ?>" data-toggle="tooltip" data-placement="bottom" title="<?php if ($v["access"] == '1') {
 									echo $tl["icons"]["i6"];
 								} else {
 									echo $tl["icons"]["i5"];
@@ -278,23 +263,17 @@ if ($page1 == "e" || $page1 == "edp" || $page1 == "ene") { ?>
 								</a>
 							</td>
 							<td>
-								<a class="btn btn-default btn-xs"
-									 href="index.php?p=user&amp;sp=password&amp;ssp=<?php echo $v["id"]; ?>"
-									 data-confirm="<?php echo $tl["user_notification"]["pass"]; ?>" data-toggle="tooltip"
-									 data-placement="bottom" title="<?php echo $tl["icons"]["i14"]; ?>">
+								<a class="btn btn-default btn-xs" href="index.php?p=user&amp;sp=password&amp;ssp=<?php echo $v["id"]; ?>" data-confirm="<?php echo $tl["user_notification"]["pass"]; ?>" data-toggle="tooltip" data-placement="bottom" title="<?php echo $tl["icons"]["i14"]; ?>">
 									<i class="fa fa-key"></i>
 								</a>
 							</td>
 							<td>
-								<a class="btn btn-default btn-xs" href="index.php?p=user&amp;sp=edit&amp;ssp=<?php echo $v["id"]; ?>"
-									 data-toggle="tooltip" data-placement="bottom" title="<?php echo $tl["icons"]["i2"]; ?>">
+								<a class="btn btn-default btn-xs" href="index.php?p=user&amp;sp=edit&amp;ssp=<?php echo $v["id"]; ?>" data-toggle="tooltip" data-placement="bottom" title="<?php echo $tl["icons"]["i2"]; ?>">
 									<i class="fa fa-edit"></i>
 								</a>
 							</td>
 							<td>
-								<a class="btn btn-default btn-xs" href="index.php?p=user&amp;sp=delete&amp;ssp=<?php echo $v["id"]; ?>"
-									 data-confirm="<?php echo $tl["user_notification"]["del"]; ?>" data-toggle="tooltip"
-									 data-placement="bottom" title="<?php echo $tl["icons"]["i1"]; ?>">
+								<a class="btn btn-default btn-xs" href="index.php?p=user&amp;sp=delete&amp;ssp=<?php echo $v["id"]; ?>" data-confirm="<?php echo $tl["user_notification"]["del"]; ?>" data-toggle="tooltip" data-placement="bottom" title="<?php echo $tl["icons"]["i1"]; ?>">
 									<i class="fa fa-trash-o"></i>
 								</a>
 							</td>
@@ -303,10 +282,9 @@ if ($page1 == "e" || $page1 == "edp" || $page1 == "ene") { ?>
 				</table>
 			</div>
 		</div>
-	</div>
-	</form>
+	</div>	</form>
 
-	<div class="col-md-12">
+	<div class="col-md-12 m-b-30">
 		<div class="icon_legend">
 			<h3><?php echo $tl["icons"]["i"]; ?></h3>
 			<i title="<?php echo $tl["icons"]["i19"]; ?>" class="fa fa-envelope-o"></i>
